Factory preference for TextDescriptor
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@2791 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/database/variable/TextDescriptor.java
@@ -743,7 +743,7 @@ public abstract class TextDescriptor implements Serializable
 		 */
 		DescriptorPref(String purpose, int initialSize)
 		{
-			cacheBits = Pref.makeLongPref("TextDescriptorFor" + purpose, prefs, swap((initialSize << Size.TXTQGRIDSH) << VTSIZESH));
+    		cacheBits = Pref.makeLongPref("TextDescriptorFor" + purpose, prefs, swap((((long)initialSize) << Size.TXTQGRIDSH) << VTSIZESH));
 			cacheColor = Pref.makeIntPref("TextDescriptorColorFor" + purpose, prefs, 0);
 			cacheFont = Pref.makeStringPref("TextDescriptorFontFor" + purpose, prefs, "");
 		}
